+++ inc/Engine/HealthCheck/HealthCheck.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Rocket\Engine\Admin;
+namespace WP_Rocket\Engine\HealthCheck;
 
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Event_Management\Subscriber_Interface;
