+++ tests/Unit/inc/Engine/HealthCheck/HealthCheck/missedCron.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\Admin\HealthCheck;
+namespace WP_Rocket\Tests\Unit\inc\Engine\HealthCheck\HealthCheck;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WP_Rocket\Admin\Options_Data;
-use WP_Rocket\Engine\Admin\HealthCheck;
+use WP_Rocket\Engine\HealthCheck\HealthCheck;
 use WPMedia\PHPUnit\Unit\TestCase;
 
 /**
- * @covers \WP_Rocket\Engine\Admin\HealthCheck::missed_cron
+ * @covers \WP_Rocket\Engine\HealthCheck\HealthCheck::missed_cron
  * @group  HealthCheck
  */
 class Test_MissedCron extends TestCase {
